+++ .env
@@ -1,5 +1,7 @@
 NEXT_PUBLIC_KAKAO_API_KEY=8f4ed639e0b0d07de0e3ce0c3365f726
 NEXT_PUBLIC_KAKAO_REST_API_KEY=265b1c201bbb3dfc32f986bccfbf8f23E
-NEXT_PUBLIC_TWITTER_ID=aXBlT0w0Zl9CYUpQZUxaemtmRkY6MTpjaQ
-NEXT_PUBLIC_TWITTER_SECRET=h6iEg2jR1_trC0TelEb50EUVEl7lAjz2HE9tZK1fD2AYryuBYY
-NEXTAUTH_URL=/api/auth
+TWITTER_CLIENT_ID=aXBlT0w0Zl9CYUpQZUxaemtmRkY6MTpjaQ
+TWITTER_CLIENT_SECRET=h6iEg2jR1_trC0TelEb50EUVEl7lAjz2HE9tZK1fD2AYryuBYY
+NEXTAUTH_SECRET=qerkasjhfkauiqeywrhkasjhfksajfh
+NEXTAUTH_URL=http://localhost:3000
+SECRET_KEY=dkasdhfjahsdfjhskdfhaksdhfakjs
