+++ next.config.js
@@ -1,7 +1,5 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  distDir: 'build',
-  output: 'export',
   webpack: (config) => {
     config.module.rules.push({
       test: /\.svg$/,
