--- src/pages/login/twitter-login.tsx
@@ -1,7 +1,0 @@
-import { useSession, signIn } from 'next-auth/react';
-
-export default function TwitterLogin() {
-  const { data: session } = useSession();
-
-  return <>{!session && <button onClick={() => signIn()}>Sign In</button>}</>;
-}
